Destroy surface when all notifications are dismissed
Fixes #18

+++ pool-buffer.c
@@ -1,5 +1,4 @@
 #define _XOPEN_SOURCE 500
-#include <assert.h>
 #include <cairo/cairo.h>
 #include <pango/pangocairo.h>
 #include <stdio.h>
@@ -49,32 +48,44 @@ static const struct wl_buffer_listener buffer_listener = {
 };
 
 static struct pool_buffer *create_buffer(struct wl_shm *shm,
-		struct pool_buffer *buf, int32_t width, int32_t height,
-		uint32_t format) {
-	uint32_t stride = width * 4;
+		struct pool_buffer *buf, int32_t width, int32_t height) {
+	const enum wl_shm_format wl_fmt = WL_SHM_FORMAT_ARGB8888;
+	const cairo_format_t cairo_fmt = CAIRO_FORMAT_ARGB32;
+
+	uint32_t stride = cairo_format_stride_for_width(cairo_fmt, width);
 	uint32_t size = stride * height;
 
-	char *name;
-	int fd = create_pool_file(size, &name);
-	assert(fd != -1);
-	void *data = mmap(NULL, size, PROT_READ | PROT_WRITE, MAP_SHARED, fd, 0);
-	struct wl_shm_pool *pool = wl_shm_create_pool(shm, fd, size);
-	buf->buffer = wl_shm_pool_create_buffer(pool, 0,
-			width, height, stride, format);
-	wl_shm_pool_destroy(pool);
-	close(fd);
-	unlink(name);
-	free(name);
-	fd = -1;
+	void *data = NULL;
+	if (size > 0) {
+		char *name;
+		int fd = create_pool_file(size, &name);
+		if (fd == -1) {
+			return NULL;
+		}
+
+		data = mmap(NULL, size, PROT_READ | PROT_WRITE, MAP_SHARED, fd, 0);
+		if (data == MAP_FAILED) {
+			return NULL;
+		}
+
+		struct wl_shm_pool *pool = wl_shm_create_pool(shm, fd, size);
+		buf->buffer =
+			wl_shm_pool_create_buffer(pool, 0, width, height, stride, wl_fmt);
+		wl_buffer_add_listener(buf->buffer, &buffer_listener, buf);
+		wl_shm_pool_destroy(pool);
+
+		close(fd);
+		fd = -1;
+		unlink(name);
+		free(name);
+	}
 
 	buf->width = width;
 	buf->height = height;
-	buf->surface = cairo_image_surface_create_for_data(data,
-			CAIRO_FORMAT_ARGB32, width, height, stride);
+	buf->surface = cairo_image_surface_create_for_data(data, cairo_fmt, width,
+		height, stride);
 	buf->cairo = cairo_create(buf->surface);
 	buf->pango = pango_cairo_create_context(buf->cairo);
-
-	wl_buffer_add_listener(buf->buffer, &buffer_listener, buf);
 	return buf;
 }
 
@@ -97,14 +108,12 @@ void destroy_buffer(struct pool_buffer *buffer) {
 struct pool_buffer *get_next_buffer(struct wl_shm *shm,
 		struct pool_buffer pool[static 2], uint32_t width, uint32_t height) {
 	struct pool_buffer *buffer = NULL;
-
 	for (size_t i = 0; i < 2; ++i) {
 		if (pool[i].busy) {
 			continue;
 		}
 		buffer = &pool[i];
 	}
-
 	if (!buffer) {
 		return NULL;
 	}
@@ -114,8 +123,7 @@ struct pool_buffer *get_next_buffer(struct wl_shm *shm,
 	}
 
 	if (!buffer->buffer) {
-		if (!create_buffer(shm, buffer, width, height,
-					WL_SHM_FORMAT_ARGB8888)) {
+		if (!create_buffer(shm, buffer, width, height)) {
 			return NULL;
 		}
 	}
